Test fixes and more polish
